+++ threedigrid_builder/interface/gridadmin.py
@@ -448,8 +448,12 @@ class GridAdminOut(OutputInterface):
             group, "line_geometries", lines.line_geometries
         )
         self.write_dataset(group, "zoom_category", lines.zoom_category)
-        self.write_dataset(group, "connection_node_end_pk", lines.line[:, 0])
-        self.write_dataset(group, "connection_node_start_pk", lines.line[:, 1])
+        self.write_dataset(
+            group, "connection_node_end_pk", lines.connection_node_start_id
+        )
+        self.write_dataset(
+            group, "connection_node_start_pk", lines.connection_node_end_id
+        )
 
         # can be collected from SQLite, but empty for now:
         self.write_dataset(group, "crest_level", fill_float)